requirements.txt
selenium~=4.9.1
webdriver-manager~=4.0.2
Appium-Python-Client~=2.10.1
requests~=2.31.0
pytest~=7.3.1
six~=1.16.0
behave~=1.2.6
jmespath~=1.0.0
setuptools~=65.5.1
simpleeval==0.9.13
# Rich reporting
allure-pytest~=2.13.2
pytest-html~=4.1.1